back-end/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pyjwt==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.1
pydantic[email]==2.7.1
